+++ crates/stitchy-core/src/files/path.rs
@@ -1,7 +1,8 @@
-
 use crate::{FileLocation, FileProperties};
-use image::{DynamicImage, ImageFormat};
+use image::{metadata::Orientation, DynamicImage, ImageFormat};
+
 use std::ffi::OsStr;
+use std::fs::File;
 use std::path::{Path, PathBuf};
 use std::time::SystemTime;
 
@@ -12,7 +13,7 @@ use std::time::SystemTime;
 pub struct FilePathWithMetadata {
     full_path: String,
     modify_time: SystemTime,
-    size_bytes: u64
+    size_bytes: u64,
 }
 
 impl Default for FilePathWithMetadata {
@@ -20,13 +21,12 @@ impl Default for FilePathWithMetadata {
         Self {
             full_path: "".to_owned(),
             modify_time: SystemTime::now(),
-            size_bytes: 0
+            size_bytes: 0,
         }
     }
 }
 
 impl FileProperties for FilePathWithMetadata {
-
     fn infer_format(&self) -> Option<ImageFormat> {
         match self.full_path.rfind('.') {
             Some(pos) => {
@@ -37,15 +37,14 @@ impl FileProperties for FilePathWithMetadata {
                     }
                 }
                 None
-            },
-            None => None
+            }
+            None => None,
         }
     }
 
     fn into_image_contents(self, print_info: bool) -> Result<DynamicImage, String> {
         let path = Path::new(&self.full_path);
-        let image = image::open(path)
-            .map_err(|_| format!("Failed to open: {:?}", path))?;
+        let image = image::open(path).map_err(|_| format!("Failed to open: {:?}", path))?;
 
         if print_info {
             if let Some(file_name) = path.file_name() {
@@ -54,7 +53,10 @@ impl FileProperties for FilePathWithMetadata {
                 println!(
                     "Path: {}, w: {}, h: {}, {}",
                     file_name.to_str().unwrap(),
-                    w, h, crate::util::make_size_string(self.size_bytes));
+                    w,
+                    h,
+                    crate::util::make_size_string(self.size_bytes)
+                );
             }
         }
 
@@ -75,12 +77,18 @@ impl FileProperties for FilePathWithMetadata {
     fn full_path(&self) -> Option<&String> {
         Some(&self.full_path)
     }
+
+    fn orientation(&self) -> Result<Orientation, String> {
+        let file = File::open(&self.full_path)
+            .map_err(|e| format!("Cannot open file {}: {:?}", &self.full_path, e))?;
+        self.decode_orientation(&file)
+    }
 }
 
 /// Wrapper for a file's location by its absolute filesystem path
 #[derive(Default, Debug)]
 pub struct FilePath {
-    path: PathBuf
+    path: PathBuf,
 }
 
 impl FilePath {
@@ -90,7 +98,6 @@ impl FilePath {
 }
 
 impl FileLocation<FilePathWithMetadata> for FilePath {
-
     fn is_file(&self) -> Result<bool, String> {
         let path_buf = PathBuf::from(&self.path);
         Ok(path_buf.is_file())
@@ -98,7 +105,8 @@ impl FileLocation<FilePathWithMetadata> for FilePath {
 
     fn extension(&self) -> Result<String, String> {
         let path_buf = PathBuf::from(&self.path);
-        path_buf.extension()
+        path_buf
+            .extension()
             .unwrap_or(OsStr::new(""))
             .to_ascii_lowercase()
             .into_string()
@@ -106,9 +114,9 @@ impl FileLocation<FilePathWithMetadata> for FilePath {
     }
 
     fn into_properties(self) -> Result<FilePathWithMetadata, String> {
-
         // Get file size and modify date from its metadata
-        let metadata = self.path
+        let metadata = self
+            .path
             .metadata()
             .map_err(|_| format!("Failed reading metadata for: {:?}", self.path))?;
         let modify_time = metadata
@@ -119,12 +127,12 @@ impl FileLocation<FilePathWithMetadata> for FilePath {
         // All seems well, get this file's properties
         let path_string = match self.path.to_str() {
             Some(s) => s.to_owned(),
-            None => return Err(format!("Failed converting to string: {:?}", self.path))
+            None => return Err(format!("Failed converting to string: {:?}", self.path)),
         };
         let properties = FilePathWithMetadata {
             full_path: path_string,
             modify_time,
-            size_bytes
+            size_bytes,
         };
         Ok(properties)
     }